Add new hook for checking Image and Linode grants; refactor accordingly
